Main control docked
Now main control could be docked
Change the tree navigation: now when an observation is checked, its
layer will be showed without show informations.
To show informations, select the observation and the details are showed
in observation info panel. To close it, press the cross on the top right
corner.
In observation info panel is possible to change the layer visibility
Bugs
On context reload, the main control undock, we must implement a variable
to store the state