tldr; Just use the same strategy as the cloud enabled in the non cloud enabled.  This means it can be populated from the application yml or via container environment variables.
The System.getenv is not an ideal solution.  There will be an annoying
workaround because using the OS env we would need to write a rule to
convert the key name.  The spring context automatically converts the
underscore, which are valid in unix systems and dot is not, and we can
retrieve them as we would expect.  Also I no longer write to the
klab.properties because it can get messy when running locally or with
persistent storage.  The -Dklab.etc should take precedent.